+++ packages/provider/lib/provider.dart
@@ -6,7 +6,6 @@ export 'src/consumer.dart';
 export 'src/inherited_provider.dart' hide autoDeferred, isWidgetTreeBuilding;
 export 'src/listenable_provider.dart';
 export 'src/provider.dart';
-export 'src/proxy_provider.dart'
-    hide NumericProxyProvider, Void, ProxyProviderBase;
+export 'src/proxy_provider.dart';
 export 'src/selector.dart';
 export 'src/value_listenable_provider.dart';