+++ packages/provider/test/proxy_provider_test.dart
@@ -3,8 +3,7 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:provider/provider.dart';
-import 'package:provider/src/proxy_provider.dart'
-    show NumericProxyProvider, Void;
+import 'package:provider/src/proxy_provider.dart' show ProxyProvider0;
 
 import 'common.dart';
 
@@ -31,6 +30,12 @@ void main() {
     builder: (context, combined, child) => combinedConsumerMock(combined),
   );
 
+  Finder findProxyProvider() {
+    return find.byWidgetPredicate(
+      (widget) => widget is InheritedProvider<Combined>,
+    );
+  }
+
   group('ProxyProvider', () {
     final combiner = CombinerMock();
     setUp(() {
@@ -44,11 +49,6 @@ void main() {
     });
     tearDown(() => clearInteractions(combiner));
 
-    Finder findProxyProvider<T>() => find.byWidgetPredicate(
-          (widget) => widget is NumericProxyProvider<T, Void, Void, Void, Void,
-              Void, Combined>,
-        );
-
     testWidgets('throws if the provided value is a Listenable/Stream',
         (tester) async {
       await tester.pumpWidget(
@@ -56,7 +56,7 @@ void main() {
           providers: [
             Provider.value(value: a),
             ProxyProvider<A, MyListenable>(
-              builder: (_, __, ___) => MyListenable(),
+              update: (_, __, ___) => MyListenable(),
             )
           ],
           child: const TextOf<MyListenable>(),
@@ -70,7 +70,7 @@ void main() {
           providers: [
             Provider.value(value: a),
             ProxyProvider<A, MyStream>(
-              builder: (_, __, ___) => MyStream(),
+              update: (_, __, ___) => MyStream(),
             )
           ],
           child: const TextOf<MyStream>(),
@@ -89,7 +89,7 @@ void main() {
           providers: [
             Provider.value(value: a),
             ProxyProvider<A, MyListenable>(
-              builder: (_, __, ___) => MyListenable(),
+              update: (_, __, ___) => MyListenable(),
             )
           ],
           child: const TextOf<MyListenable>(),
@@ -101,7 +101,7 @@ void main() {
           providers: [
             Provider.value(value: a),
             ProxyProvider<A, MyStream>(
-              builder: (_, __, ___) => MyStream(),
+              update: (_, __, ___) => MyStream(),
             )
           ],
           child: const TextOf<MyStream>(),
@@ -109,30 +109,27 @@ void main() {
       );
     });
 
-    testWidgets('initialBuilder creates initial value', (tester) async {
-      final initialBuilder = InitialValueBuilderMock<Combined>();
-      final key = GlobalKey();
+    testWidgets('create creates initial value', (tester) async {
+      final create = InitialValueBuilderMock<Combined>();
 
-      when(initialBuilder(any)).thenReturn(Combined(null, null, null));
+      when(create(any)).thenReturn(Combined(null, null, null));
 
       await tester.pumpWidget(
         MultiProvider(
           providers: [
             Provider.value(value: a),
             ProxyProvider<A, Combined>(
-              key: key,
-              initialBuilder: initialBuilder,
-              builder: combiner,
+              create: create,
+              update: combiner,
             )
           ],
           child: mockConsumer,
         ),
       );
 
-      final details = verify(initialBuilder(captureAny))..called(1);
-      expect(details.captured.first, equals(key.currentContext));
+      verify(create(argThat(isNotNull)))..called(1);
 
-      verify(combiner(key.currentContext, a, Combined(null, null, null)));
+      verify(combiner(argThat(isNotNull), a, Combined(null, null, null)));
     });
     testWidgets('consume another providers', (tester) async {
       await tester.pumpWidget(
@@ -140,14 +137,14 @@ void main() {
           providers: [
             Provider.value(value: a),
             ProxyProvider<A, Combined>(
-              builder: combiner,
+              update: combiner,
             )
           ],
           child: mockConsumer,
         ),
       );
 
-      final context = tester.element(findProxyProvider<A>());
+      final context = tester.element(findProxyProvider());
 
       verify(combinedConsumerMock(Combined(context, null, a))).called(1);
       verifyNoMoreInteractions(combinedConsumerMock);
@@ -156,9 +153,21 @@ void main() {
       verifyNoMoreInteractions(combiner);
     });
 
-    test('throws if builder is null', () {
+    test('throws if update is null', () {
       // ignore: prefer_const_constructors
-      expect(() => ProxyProvider<A, Combined>(builder: null),
+      expect(
+          () => ProxyProvider<A, Combined>(update: null), throwsAssertionError);
+      expect(
+          () => ProxyProvider0<Combined>(update: null), throwsAssertionError);
+      expect(() => ProxyProvider2<A, B, Combined>(update: null),
+          throwsAssertionError);
+      expect(() => ProxyProvider3<A, B, C, Combined>(update: null),
+          throwsAssertionError);
+      expect(() => ProxyProvider4<A, B, C, D, Combined>(update: null),
+          throwsAssertionError);
+      expect(() => ProxyProvider5<A, B, C, D, E, Combined>(update: null),
+          throwsAssertionError);
+      expect(() => ProxyProvider6<A, B, C, D, E, F, Combined>(update: null),
           throwsAssertionError);
     });
 
@@ -168,7 +177,7 @@ void main() {
           providers: [
             Provider.value(value: a),
             ProxyProvider<A, Combined>(
-              builder: combiner,
+              update: combiner,
             )
           ],
           child: mockConsumer,
@@ -182,13 +191,13 @@ void main() {
           providers: [
             Provider.value(value: a2),
             ProxyProvider<A, Combined>(
-              builder: combiner,
+              update: combiner,
             )
           ],
           child: mockConsumer,
         ),
       );
-      final context = tester.element(findProxyProvider<A>());
+      final context = tester.element(findProxyProvider());
 
       verifyInOrder([
         combiner(context, a, null),
@@ -209,7 +218,7 @@ void main() {
         MultiProvider(
           providers: [
             Provider.value(value: a),
-            ProxyProvider<A, Combined>(builder: combiner, dispose: dispose)
+            ProxyProvider<A, Combined>(update: combiner, dispose: dispose)
           ],
           child: mockConsumer,
         ),
@@ -222,18 +231,23 @@ void main() {
         MultiProvider(
           providers: [
             Provider.value(value: a2),
-            ProxyProvider<A, Combined>(builder: combiner, dispose: dispose2)
+            ProxyProvider<A, Combined>(update: combiner, dispose: dispose2)
           ],
           child: mockConsumer,
         ),
       );
-      final context = tester.element(findProxyProvider<A>());
+      final context = tester.element(findProxyProvider());
+
+      verify(
+        dispose(context, Combined(context, null, a)),
+      );
 
       await tester.pumpWidget(Container());
 
-      verifyZeroInteractions(dispose);
       verify(
-          dispose2(context, Combined(context, Combined(context, null, a), a2)));
+        dispose2(context, Combined(context, Combined(context, null, a), a2)),
+      );
+      verifyNoMoreInteractions(dispose);
     });
 
     testWidgets("don't rebuild descendants if value doesn't change",
@@ -243,7 +257,7 @@ void main() {
           providers: [
             Provider.value(value: a),
             ProxyProvider<A, Combined>(
-              builder: (c, a, p) => combiner(c, a, null),
+              update: (c, a, p) => combiner(c, a, null),
             )
           ],
           child: mockConsumer,
@@ -258,7 +272,7 @@ void main() {
               updateShouldNotify: (A _, A __) => true,
             ),
             ProxyProvider<A, Combined>(
-              builder: (c, a, p) {
+              update: (c, a, p) {
                 combiner(c, a, p);
                 return p;
               },
@@ -267,7 +281,7 @@ void main() {
           child: mockConsumer,
         ),
       );
-      final context = tester.element(findProxyProvider<A>());
+      final context = tester.element(findProxyProvider());
 
       verifyInOrder([
         combiner(context, a, null),
@@ -298,7 +312,7 @@ void main() {
           Provider<String>.value(
               value: 'Hello', updateShouldNotify: (_, __) => true),
           ProxyProvider<String, String>(
-            builder: (_, value, __) => value,
+            update: (_, value, __) => value,
             updateShouldNotify: shouldNotify,
           ),
         ],
@@ -310,7 +324,7 @@ void main() {
           Provider<String>.value(
               value: 'Hello', updateShouldNotify: (_, __) => true),
           ProxyProvider<String, String>(
-            builder: (_, value, __) => value,
+            update: (_, value, __) => value,
             updateShouldNotify: shouldNotify,
           ),
         ],
@@ -329,47 +343,30 @@ void main() {
       await tester.pumpWidget(MultiProvider(
         providers: [
           Provider.value(value: a),
-          ProxyProvider<A, Combined>(builder: (c, a, p) => Combined(c, p, a)),
+          ProxyProvider<A, Combined>(update: (c, a, p) => Combined(c, p, a)),
         ],
         child: Container(key: key),
       ));
-      final context = tester.element(findProxyProvider<A>());
+      final context = tester.element(findProxyProvider());
 
       expect(
         Provider.of<Combined>(key.currentContext),
         Combined(context, null, a),
       );
     });
-    test('works with MultiProvider #2', () {
-      final provider = ProxyProvider<A, B>(
-        key: const Key('42'),
-        initialBuilder: (_) => null,
-        builder: (_, __, ___) => null,
-        updateShouldNotify: (_, __) => null,
-        dispose: (_, __) {},
-        child: Container(),
-      );
-      var child2 = Container();
-      final clone = provider.cloneWithChild(child2);
-
-      expect(clone.child, equals(child2));
-      expect(clone.key, equals(provider.key));
-      expect(clone.initialBuilder, equals(provider.initialBuilder));
-      expect(clone.builder, equals(provider.builder));
-      expect(clone.updateShouldNotify, equals(provider.updateShouldNotify));
-      expect(clone.dispose, equals(provider.dispose));
-      // expect(clone.providerBuilder, equals(provider.providerBuilder));
-    });
 
     // useful for libraries such as Mobx where events are synchronously
     // dispatched
     testWidgets(
-        'builder callback can trigger descendants setState synchronously',
+        'update callback can trigger descendants setState synchronously',
         (tester) async {
       var statefulBuildCount = 0;
       void Function(VoidCallback) setState;
 
-      final statefulBuilder = StatefulBuilder(builder: (_, s) {
+      final statefulBuilder = StatefulBuilder(builder: (context, s) {
+        // force update to be called
+        Provider.of<Combined>(context, listen: false);
+
         setState = s;
         statefulBuildCount++;
         return Container();
@@ -378,15 +375,21 @@ void main() {
       await tester.pumpWidget(MultiProvider(
         providers: [
           Provider.value(value: a),
-          ProxyProvider<A, Combined>(builder: (c, a, p) => Combined(c, p, a)),
+          ProxyProvider<A, Combined>(update: (c, a, p) => Combined(c, p, a)),
         ],
         child: statefulBuilder,
       ));
 
+      expect(
+        statefulBuildCount,
+        1,
+        reason: 'update must not be called asynchronously',
+      );
+
       await tester.pumpWidget(MultiProvider(
         providers: [
           Provider.value(value: A()),
-          ProxyProvider<A, Combined>(builder: (c, a, p) {
+          ProxyProvider<A, Combined>(update: (c, a, p) {
             setState(() {});
             return Combined(c, p, a);
           }),
@@ -397,16 +400,12 @@ void main() {
       expect(
         statefulBuildCount,
         2,
-        reason: 'builder must not be called asynchronously',
+        reason: 'update must not be called asynchronously',
       );
     });
   });
 
   group('ProxyProvider variants', () {
-    Finder findProxyProvider<A, B, C, D, E, F>() => find.byWidgetPredicate(
-          (widget) =>
-              widget is NumericProxyProvider<A, B, C, D, E, F, Combined>,
-        );
     testWidgets('ProxyProvider2', (tester) async {
       await tester.pumpWidget(
         MultiProvider(
@@ -418,8 +417,8 @@ void main() {
             Provider.value(value: e),
             Provider.value(value: f),
             ProxyProvider2<A, B, Combined>(
-              initialBuilder: (_) => Combined(null, null, null),
-              builder: (context, a, b, previous) =>
+              create: (_) => Combined(null, null, null),
+              update: (context, a, b, previous) =>
                   Combined(context, previous, a, b),
             )
           ],
@@ -427,8 +426,7 @@ void main() {
         ),
       );
 
-      final context =
-          tester.element(findProxyProvider<A, B, Void, Void, Void, Void>());
+      final context = tester.element(findProxyProvider());
 
       verify(
         combinedConsumerMock(
@@ -447,8 +445,8 @@ void main() {
             Provider.value(value: e),
             Provider.value(value: f),
             ProxyProvider3<A, B, C, Combined>(
-              initialBuilder: (_) => Combined(null, null, null),
-              builder: (context, a, b, c, previous) =>
+              create: (_) => Combined(null, null, null),
+              update: (context, a, b, c, previous) =>
                   Combined(context, previous, a, b, c),
             )
           ],
@@ -456,8 +454,7 @@ void main() {
         ),
       );
 
-      final context =
-          tester.element(findProxyProvider<A, B, C, Void, Void, Void>());
+      final context = tester.element(findProxyProvider());
 
       verify(
         combinedConsumerMock(
@@ -476,8 +473,8 @@ void main() {
             Provider.value(value: e),
             Provider.value(value: f),
             ProxyProvider4<A, B, C, D, Combined>(
-              initialBuilder: (_) => Combined(null, null, null),
-              builder: (context, a, b, c, d, previous) =>
+              create: (_) => Combined(null, null, null),
+              update: (context, a, b, c, d, previous) =>
                   Combined(context, previous, a, b, c, d),
             )
           ],
@@ -485,8 +482,7 @@ void main() {
         ),
       );
 
-      final context =
-          tester.element(findProxyProvider<A, B, C, D, Void, Void>());
+      final context = tester.element(findProxyProvider());
 
       verify(
         combinedConsumerMock(
@@ -505,8 +501,8 @@ void main() {
             Provider.value(value: e),
             Provider.value(value: f),
             ProxyProvider5<A, B, C, D, E, Combined>(
-              initialBuilder: (_) => Combined(null, null, null),
-              builder: (context, a, b, c, d, e, previous) =>
+              create: (_) => Combined(null, null, null),
+              update: (context, a, b, c, d, e, previous) =>
                   Combined(context, previous, a, b, c, d, e),
             )
           ],
@@ -514,7 +510,7 @@ void main() {
         ),
       );
 
-      final context = tester.element(findProxyProvider<A, B, C, D, E, Void>());
+      final context = tester.element(findProxyProvider());
 
       verify(
         combinedConsumerMock(
@@ -533,8 +529,8 @@ void main() {
             Provider.value(value: e),
             Provider.value(value: f),
             ProxyProvider6<A, B, C, D, E, F, Combined>(
-              initialBuilder: (_) => Combined(null, null, null),
-              builder: (context, a, b, c, d, e, f, previous) =>
+              create: (_) => Combined(null, null, null),
+              update: (context, a, b, c, d, e, f, previous) =>
                   Combined(context, previous, a, b, c, d, e, f),
             )
           ],
@@ -542,7 +538,7 @@ void main() {
         ),
       );
 
-      final context = tester.element(findProxyProvider<A, B, C, D, E, F>());
+      final context = tester.element(findProxyProvider());
 
       verify(
         combinedConsumerMock(